requirements.txt
argcomplete==1.9.2
base58==0.2.4
boto3==1.7.25
botocore==1.10.25
certifi==2018.4.16
cfn-flip==1.0.3
chardet==3.0.4
click==6.7
Django==2.0.5
django-split-settings==0.3.0
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
docutils==0.14
durationpy==0.5
future==0.16.0
hjson==3.0.1
idna==2.6
jmespath==0.9.3
kappa==0.6.0
lambda-packages==0.19.0
placebo==0.8.1
PyJWT==1.6.3
python-dateutil==2.6.1
python-slugify==1.2.4
pytz==2018.4
PyYAML==3.12
requests==2.18.4
s3transfer==0.1.13
six==1.11.0
toml==0.9.4
tqdm==4.19.1
troposphere==2.2.1
Unidecode==1.0.22
urllib3==1.22
Werkzeug==0.12
wsgi-request-logger==0.4.6
zappa==0.45.1
